projecthub/requirements.txt
asgiref==3.9.1
boto3==1.40.12
botocore==1.40.12
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.3
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
jmespath==1.0.1
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.16.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0